Start switching away from flocker-reportstate usage in acceptance tests.
